Revert "Attempt to move item list state to listview"
This reverts commit 1d652af4c2360fc2fb6a4afe8eb7a002f1731618.
